[CHORE] 용량 데이터 타입 변경

+++ src/main/java/com/tiki/server/document/dto/response/DocumentsGetResponse.java
@@ -26,7 +26,7 @@ public record DocumentsGetResponse(
 		@NotNull long documentId,
 		@NotNull String name,
 		@NotNull String url,
-		@NotNull double capacity,
+		@NotNull long capacity,
 		@NotNull LocalDateTime createdTime
 	) {
