+++ src/main/java/com/tiki/server/team/dto/response/UsageGetResponse.java
@@ -7,11 +7,11 @@ import lombok.Builder;
 
 @Builder(access = PRIVATE)
 public record UsageGetResponse(
-	@NotNull double capacity,
-	@NotNull double usage
+	@NotNull long capacity,
+	@NotNull long usage
 ) {
 
-	public static UsageGetResponse of(final double capacity, final double usage) {
+	public static UsageGetResponse of(final long capacity, final long usage) {
 		return UsageGetResponse.builder()
 			.capacity(capacity)
 			.usage(usage)
